requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
ansiwrap==0.8.4
arrow==0.15.5
asn1crypto==1.2.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.8.2
binaryornot==0.4.4
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
blis==0.2.4
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.0
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.3
conda==4.8.2
conda-package-handling==1.6.0
configparser==4.0.2
confuse==1.0.0
contextlib2==0.6.0.post1
cookiecutter==1.7.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
Cython==0.29.14
cytoolz==0.10.1
dask==2.9.2
dataclasses==0.6
datalab==1.1.5
decorator==4.4.1
defusedxml==0.6.0
distributed==2.9.3
docker==4.1.0
docutils==0.16
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
fairing==0.5.3
fastai==1.0.60
fastcache==1.1.0
fastprogress==0.2.2
filelock==3.0.12
Flask==1.1.1
fsspec==0.6.2
future==0.18.2
gast==0.3.3
gcsfs==0.6.0
gevent==1.4.0
gitdb2==2.0.6
GitPython==3.0.5
glob2==0.7
gmpy2==2.0.8
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.23.1
google-cloud-core==1.2.0
google-cloud-dataproc==0.6.1
google-cloud-datastore==1.10.0
google-cloud-language==1.3.0
google-cloud-logging==1.14.0
google-cloud-monitoring==0.31.1
google-cloud-spanner==1.13.0
google-cloud-storage==1.25.0
google-cloud-translate==2.0.0
google-pasta==0.1.8
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.17.0
hypothesis==4.54.2
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.4.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
jeepney==0.4.2
Jinja2==2.10.3
jinja2-time==0.2.0
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-aihub-deploy-extension==0.1
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-http-over-ws==0.0.7
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.6
jupyterlab-git==0.9.0
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==20.0.0
kiwisolver==1.1.0
kubernetes==10.0.1
lazy-object-proxy==1.4.3
libarchive-c==2.8
lief==0.9.0
llvmlite==0.31.0
locket==0.2.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==8.0.2
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
murmurhash==1.0.2
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.6.1
nbdime==1.1.0
nbformat==5.0.4
nbpresent==3.0.2
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
notebook-executor==0.2
numba==0.47.0
numexpr==2.7.0
numpy==1.18.1
numpydoc==0.9.2
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.3
packaging==20.1
pandas==0.25.3
pandas-profiling==2.4.0
pandocfilters==1.4.2
papermill==1.2.1
parso==0.6.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
phik==0.9.8
pickleshare==0.7.5
Pillow==6.2.0
Pillow-SIMD==6.2.2.post1
pkginfo==1.5.0.1
plac==0.9.6
plotly==4.5.0
pluggy==0.13.1
ply==3.11
poyo==0.5.0
preshed==2.0.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.4
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.4
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-pylint==0.14.1
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.3.2
seaborn==0.9.0
SecretStorage==3.1.2
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.1.8
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.13
sqlparse==0.3.0
srsly==0.1.0
statsmodels==0.10.1
sympy==1.5.1
tables==3.6.1
tb-nightly==1.14.0a20190301
tblib==1.6.0
tenacity==6.0.0
tensorflow-gpu==2.0.0a0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textwrap3==0.9.2
tf-estimator-nightly==1.14.0.dev2019030115
thinc==7.0.8
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
tqdm==4.42.0
traitlets==4.3.3
typed-ast==1.4.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.24.2
virtualenv==16.7.9
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
